requirements/test.txt
-r base.txt

atomicwrites==1.1.5
attrs==18.1.0
bumpversion==0.5.3
coverage==4.5.1
flake8==3.5.0
mccabe==0.6.1
mock==2.0.0
more-itertools==4.3.0
packaging==17.1
pbr==4.2.0
pluggy==0.7.1
py==1.5.4
pycodestyle==2.3.1
pyflakes==1.6.0
pyparsing==2.2.0
pytest-cov==2.5.1
pytest-django==3.3.3
pytest==3.7.0
selenium==3.14.0
six==1.11.0
tox==3.1.2
urllib3==1.23
virtualenv==16.0.0